Rename modules rust::parser, rust::serializer to decoder, parser
