+++ pdl-compiler/src/backends/rust/encoder.rs
